feat: enable moderator quiz launch
