+++ src/i18n/dictionaries/sk.json
@@ -67,7 +67,7 @@
         "name": "Herd Vote – tímový kvíz",
         "description": "Otestujte, ako dobre sa poznáte. Zábavné otázky od ľahkých až po nečakané – kto vás pozná lepšie?",
         "cta": "Začať kvíz",
-        "link": "/apps/herd-vote",
+        "link": "/apps/quiz",
         "manual": [
           "Vytvorte alebo sa pripojte k hre",
           "Odpovedajte na otázky",
